feat: migrate Votes.spec
